More fiddling with the Spotlight opener. Fix a crash when generating thumbnail images and the window closes.

+++ Classes/Managed Objects/TSSTManagedGroup.m
@@ -357,14 +357,14 @@
     
     NSFileManager * fileManager = [NSFileManager defaultManager];
 	NSData * fileData;
-	int collision = 0;
+	NSInteger collision = 0;
     NSString * archivePath = nil;
 	NSInteger counter, archivedFilesCount = [imageArchive numberOfEntries];
 	NSError * error;
 	if([imageArchive isSolid])
 	{
 		do {
-			archivePath = [NSString stringWithFormat: @"SC-images-%i", collision];
+			archivePath = [NSString stringWithFormat: @"SC-images-%li", (long)collision];
 			archivePath = [NSTemporaryDirectory() stringByAppendingPathComponent: archivePath];
 			++collision;
 		} while (![fileManager createDirectoryAtPath: archivePath withIntermediateDirectories: YES attributes: nil error: &error]);
@@ -394,7 +394,7 @@
 				
                 collision = 0;
                 do {
-                    archivePath = [NSString stringWithFormat: @"%i-%@", collision, fileName];
+                    archivePath = [NSString stringWithFormat: @"%li-%@", (long)collision, fileName];
                     archivePath = [NSTemporaryDirectory() stringByAppendingPathComponent: archivePath];
                     ++collision;
                 } while ([fileManager fileExistsAtPath: archivePath]);
